+++ CHANGELOG.md
@@ -34,6 +34,7 @@ All notable changes to this project will be documented in this file.<br>
 - Added new functions in NSAttributedString extension to handle left, right, center, justified and natural text alignment
 - Added `smoothLinearGradient(colors:, direction:)` function in UIView extension that uses `CGGradient` instead of `CAGradientLayer`
 - Added iPad Pro 12.9 2nd Generation and iPad Pro 10.5 devices in UIDevice extension
+- Added `fade(label:, duration: = default, toString:)` and `fade(label:, duration: = default, toAttributedString:)` animation functions in WKInterfaceController extension
 
 ### Improved
 - Now `dateString(format: = default)` in Date extension has a new `locale` parameter with a `default` value, so no changes are required